requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Data collection - ALL FREE, NO API KEYS
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2

# AI/ML
groq==0.4.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
scipy==1.11.4

# Data processing
pandas==2.1.3
numpy==1.26.2

# Visualization
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
